Add the ability to nuke user data outright #automation (#416)
* cmd/frontend/db: on user deletion, also soft delete discussions data (fixes sourcegraph/sourcegraph#412)

* cmd/frontend/db: add support for hard user data deletion (helps sourcegraph/enterprise#13602)

* cmd/frontend/graphqlbackend: expose hard user data deletion (helps sourcegraph/enterprise#13602)

* add the ability to nuke users (for e.g. GDPR-style user data deletion requests)

Helps sourcegraph/enterprise#13602

+++ cmd/frontend/graphqlbackend/site_admin.go
@@ -11,6 +11,7 @@ import (
 
 func (*schemaResolver) DeleteUser(ctx context.Context, args *struct {
 	User graphql.ID
+	Hard *bool
 }) (*EmptyResponse, error) {
 	// 🚨 SECURITY: Only site admins can delete users.
 	if err := backend.CheckCurrentUserIsSiteAdmin(ctx); err != nil {
@@ -30,8 +31,14 @@ func (*schemaResolver) DeleteUser(ctx context.Context, args *struct {
 		return nil, errors.New("unable to delete current user")
 	}
 
-	if err := db.Users.Delete(ctx, userID); err != nil {
-		return nil, err
+	if args.Hard != nil && *args.Hard {
+		if err := db.Users.HardDelete(ctx, userID); err != nil {
+			return nil, err
+		}
+	} else {
+		if err := db.Users.Delete(ctx, userID); err != nil {
+			return nil, err
+		}
 	}
 	return &EmptyResponse{}, nil
 }
